+++ src/util/Tile.java
@@ -1,7 +1,7 @@
 package util;
 
 public class Tile extends VTile<String> {
-    public Tile(int row, int col, String value) {
+    public Tile(long row, long col, String value) {
         super(row, col, value);
     }
 }
